refactor(ui): rewire internal imports and exports map for taxonomy paths
