+++ apps/packages/ui/src/icons/command-palette-icons.tsx
@@ -1,4 +1,4 @@
-import type { IconProps } from "./icons/types";
+import type { IconProps } from "./types";
 
 export type CommandPaletteGlyphName =
   | "arrow-left"
