+++ apps/packages/ui/src/icons/index.tsx
@@ -1,9 +1,9 @@
-// Compatibility entry for the existing @proliferate/ui/icons subpath.
+// Barrel entry for the @proliferate/ui/icons subpath.
 // New ui-internal code should import the owned modules under ./icons/* directly.
-export type { IconProps } from "./icons/types";
-export { ChevronRight, ChevronRightActivity, ChevronDown, ChevronUpDown, ChevronDownUp, Minus, Plus, MiniPlus, X, Copy, Search, SearchActivity, Hash, Keyboard, KeyRound, Pencil, Trash, Archive, Clock, Settings, SlidersHorizontal, Palette, CircleQuestion, MessageCircleQuestion, LifeBuoy, ArrowLeft, ArrowRight, LogOut, ArrowUp, ArrowDown, List, ListFilter, StopSquare, Play, Pause, ArrowUpRight, ExternalLink, Check, XLines, Fork, Undo, Link2, WrapText, MoreHorizontal, GripVertical, SendIcon, History, RotateCcw } from "./icons/core";
-export { GitMerge, GitBranch, GitPullRequest, Terminal, SquareTerminal, CommandWindow, ReadBook, FileText, FileIcon, InlinePathMentionIcon, FileCode, StackedFiles, FilePen, FileDiff, FilePenActivity, FilePlus, FolderPlus, FolderList, FolderTree, SquarePlus, Columns2, Folder, FolderOpen, FolderClosedFilled, FolderPlusFilled, FolderFilled, FolderRemote, PrBranchGlyph, PrMergedGlyph, Filter, CollapseAll, ExpandAll, GitCommit, Tree, GitBranchIcon, ProjectNotebook } from "./icons/workspace";
-export { PixelAgentSprite, Robot, ClipboardList, ClipboardListFilled, ReadModeFilled, EditModeFilled, BuildModeFilled, OpencodeBuildModeFilled, OpencodePlanModeFilled, MessageSquareFilled, MessageSquarePlus, AddMessage, AddPlan, Brain, ThinkingGlyph, PlanningIcon, Shield, MessageSquare, StickyNote, ScratchPadIcon, Zap, Sparkles, ClaudeSparkle, BotMessageSquare, Target } from "./icons/product";
-export { Mail, Calendar, CalendarClock, CreditCard, Discord, GitHub, Globe, Smartphone, Sun, Moon, Monitor, Home, Grid, CloudUpload, Blocks, CircleUser, Building2, UsersRound, CloudIcon, Server, RefreshCw } from "./icons/platform";
-export { CheckCircleFilled, Circle, CircleAlert, ShieldCheckFilled, Spinner, BrailleSweepBadge } from "./icons/status";
-export { AppShellReviewIcon, AppShellBrowserIcon, AppShellTerminalIcon, AppShellPlusIcon, AppShellPanelToggleIcon, AppShellTabCloseIcon, SplitPanel, SplitPanelLeft, SplitPanelRight } from "./icons/app-shell";
+export type { IconProps } from "./types";
+export { ChevronRight, ChevronRightActivity, ChevronDown, ChevronUpDown, ChevronDownUp, Minus, Plus, MiniPlus, X, Copy, Search, SearchActivity, Hash, Keyboard, KeyRound, Pencil, Trash, Archive, Clock, Settings, SlidersHorizontal, Palette, CircleQuestion, MessageCircleQuestion, LifeBuoy, ArrowLeft, ArrowRight, LogOut, ArrowUp, ArrowDown, List, ListFilter, StopSquare, Play, Pause, ArrowUpRight, ExternalLink, Check, XLines, Fork, Undo, Link2, WrapText, MoreHorizontal, GripVertical, SendIcon, History, RotateCcw } from "./core";
+export { GitMerge, GitBranch, GitPullRequest, Terminal, SquareTerminal, CommandWindow, ReadBook, FileText, FileIcon, InlinePathMentionIcon, FileCode, StackedFiles, FilePen, FileDiff, FilePenActivity, FilePlus, FolderPlus, FolderList, FolderTree, SquarePlus, Columns2, Folder, FolderOpen, FolderClosedFilled, FolderPlusFilled, FolderFilled, FolderRemote, PrBranchGlyph, PrMergedGlyph, Filter, CollapseAll, ExpandAll, GitCommit, Tree, GitBranchIcon, ProjectNotebook } from "./workspace";
+export { PixelAgentSprite, Robot, ClipboardList, ClipboardListFilled, ReadModeFilled, EditModeFilled, BuildModeFilled, OpencodeBuildModeFilled, OpencodePlanModeFilled, MessageSquareFilled, MessageSquarePlus, AddMessage, AddPlan, Brain, ThinkingGlyph, PlanningIcon, Shield, MessageSquare, StickyNote, ScratchPadIcon, Zap, Sparkles, ClaudeSparkle, BotMessageSquare, Target } from "./product";
+export { Mail, Calendar, CalendarClock, CreditCard, Discord, GitHub, Globe, Smartphone, Sun, Moon, Monitor, Home, Grid, CloudUpload, Blocks, CircleUser, Building2, UsersRound, CloudIcon, Server, RefreshCw } from "./platform";
+export { CheckCircleFilled, Circle, CircleAlert, ShieldCheckFilled, Spinner, BrailleSweepBadge } from "./status";
+export { AppShellReviewIcon, AppShellBrowserIcon, AppShellTerminalIcon, AppShellPlusIcon, AppShellPanelToggleIcon, AppShellTabCloseIcon, SplitPanel, SplitPanelLeft, SplitPanelRight } from "./app-shell";
