+++ apps/packages/ui/src/patterns/ComposerControlButton.tsx
@@ -1,5 +1,5 @@
 import { forwardRef, type ButtonHTMLAttributes, type ReactNode } from "react";
-import { Button } from "./Button";
+import { Button } from "../primitives/Button";
 
 interface ComposerControlButtonProps extends Omit<ButtonHTMLAttributes<HTMLButtonElement>, "children"> {
   icon?: ReactNode;
